separate better concerns when creating update or upgrade object
